[FIX] ChangePassword, Follow, Bookmark
