+++ README.md
@@ -29,3 +29,11 @@
   - 비밀번호 변경 VIEW 구축(ing)
 ```
 
+## 23.8.30.
+```
+- Account
+  - 비밀번호 변경 VIEW 구축 완료
+  - 팔로우, 북마크 코드 수정(간결화)
+  - views 폴더 정리(가독성을 위해 파일 분리)
+  - serializers.py 정리
+```
