+++ backend/accounts/views/__init__.py
@@ -1,2 +1,5 @@
-from .social_views import *
-from .views import *
+from .social_account_views import *
+from .account_views import *
+from .bookmark_views import *
+from .follow_views import *
+from .etc_views import *
